Hyperdrive/projects/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.32
psycopg[binary]==3.2.1
pydantic==2.8.2
pydantic-settings==2.3.4
python-jose==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
redis==5.0.8
bleach==6.1.0
email-validator==2.2.0
python-slugify==8.0.4